move document related emails to separate mailer
