+++ app/mailers/operator_mailer.rb
@@ -3,22 +3,6 @@ class OperatorMailer < ApplicationMailer
 
   helper :date
 
-  def expiring_documents(operator, user, documents)
-    @documents = documents.sort_by { |d| [d.fmu&.name || "_", d.required_operator_document.name] }
-    @user = user
-    @operator = operator
-
-    mail to: user.email, subject: I18n.t("operator_mailer.expiring_documents.subject")
-  end
-
-  def expired_documents(operator, user, documents)
-    @documents = documents.sort_by { |d| [d.fmu&.name || "_", d.required_operator_document.name] }
-    @user = user
-    @operator = operator
-
-    mail to: user.email, subject: I18n.t("operator_mailer.expired_documents.subject", count: @documents.count)
-  end
-
   # An email that contains the a quarterly report of an operator
   # It lists:
   # 1. Current transparency score
@@ -49,19 +33,5 @@ class OperatorMailer < ApplicationMailer
     end
 
     mail to: user.email, subject: I18n.t("operator_mailer.quarterly_newsletter.subject", company: operator.name)
-  end
-
-  def document_valid(document, user)
-    @operator = document.operator
-    @document = document
-    @user = user
-    mail to: user.email, subject: I18n.t("operator_mailer.document_valid.subject")
-  end
-
-  def document_invalid(document, user)
-    @operator = document.operator
-    @document = document
-    @user = user
-    mail to: user.email, subject: I18n.t("operator_mailer.document_invalid.subject")
   end
 end
