+++ app/views/mailers/operator_document_mailer/document_invalid.text.erb
