+++ app/views/mailers/operator_document_mailer/document_valid.text.erb
